fix: re organize pulse/react code outside of the tickets dir
