+++ src/features/media/PulsePage.tsx
